[jit] add reuse_stubs_fn option to create_script_module to allow reuse rule for submodule
when compiling submodule, we always use the default method inference rule, which includes `forward` and methods it calls, this is also the case for tracing. For tracing we don't want to compile the forward by default for submodule. This PR allows the recursive scripting to reuse stubs_fn to create its submodule instead of alwaysing using the default rule.

Fixes https://github.com/pytorch/pytorch/issues/43729

ghstack-source-id: 2201f7b311e08efb76c8ccd6ec989aaa5b8049eb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43872